refactor(header-chain): package validation by phase
