--- crates/zakura-header-chain/src/validation/contextual.rs
@@ -1,963 +1,0 @@
-//! Block difficulty adjustment calculations for contextual validation.
-//!
-//! The contextual validator calculates the following consensus rules:
-//!  * `ThresholdBits` from the Zcash Specification,
-//!  * the Testnet minimum difficulty adjustment from ZIPs 205 and 208, and
-//!  * `median-time-past`.
-
-use std::cmp::{max, min};
-
-use chrono::{DateTime, Duration, Utc};
-use thiserror::Error;
-
-use zakura_chain::{
-    block::{self, Block},
-    parameters::{Network, NetworkUpgrade, POW_AVERAGING_WINDOW},
-    work::difficulty::{CompactDifficulty, ExpandedDifficulty, ParameterDifficulty as _, U256},
-    BoundedVec,
-};
-
-/// The median block span for time median calculations.
-///
-/// `PoWMedianBlockSpan` in the Zcash specification.
-pub const POW_MEDIAN_BLOCK_SPAN: usize = 11;
-
-/// The overall block span used for adjusting Zcash block difficulty.
-///
-/// `PoWAveragingWindow + PoWMedianBlockSpan` in the Zcash specification based on
-/// > ActualTimespan(height : N) := MedianTime(height) − MedianTime(height − PoWAveragingWindow)
-pub const POW_ADJUSTMENT_BLOCK_SPAN: usize = POW_AVERAGING_WINDOW + POW_MEDIAN_BLOCK_SPAN;
-
-/// Durable predecessors needed below a separately retained parent frontier.
-pub const POW_PREDECESSOR_CONTEXT_SPAN: usize = POW_ADJUSTMENT_BLOCK_SPAN - 1;
-
-/// The damping factor for median timespan variance.
-///
-/// `PoWDampingFactor` in the Zcash specification.
-pub const POW_DAMPING_FACTOR: i32 = 4;
-
-/// The maximum upward adjustment percentage for median timespan variance.
-///
-/// `PoWMaxAdjustUp * 100` in the Zcash specification.
-pub const POW_MAX_ADJUST_UP_PERCENT: i32 = 16;
-
-/// The maximum downward adjustment percentage for median timespan variance.
-///
-/// `PoWMaxAdjustDown * 100` in the Zcash specification.
-pub const POW_MAX_ADJUST_DOWN_PERCENT: i32 = 32;
-
-/// The maximum number of seconds between the `median-time-past` of a block,
-/// and the block's `time` field.
-///
-/// Part of the block header consensus rules in the Zcash specification.
-pub const BLOCK_MAX_TIME_SINCE_MEDIAN: u32 = 90 * 60;
-
-/// The difficulty context calculates a block's adjusted difficulty.
-pub struct AdjustedDifficulty {
-    /// The `header.time` field from the candidate block
-    candidate_time: DateTime<Utc>,
-    /// The coinbase height from the candidate block
-    ///
-    /// Header validation calculates this field from the previous block height.
-    candidate_height: block::Height,
-    /// The configured network
-    network: Network,
-    /// The `header.difficulty_threshold`s from the previous
-    /// `PoWAveragingWindow + PoWMedianBlockSpan` (28) blocks, in reverse height
-    /// order.
-    relevant_difficulty_thresholds: BoundedVec<CompactDifficulty, 1, POW_ADJUSTMENT_BLOCK_SPAN>,
-    /// The `header.time`s from the previous
-    /// `PoWAveragingWindow + PoWMedianBlockSpan` (28) blocks, in reverse height
-    /// order.
-    ///
-    /// The calculation uses only the first and last `PoWMedianBlockSpan` times.
-    /// The calculation ignores times `11..=16`.
-    relevant_times: BoundedVec<DateTime<Utc>, 1, POW_ADJUSTMENT_BLOCK_SPAN>,
-}
-
-/// Invalid branch context supplied to a difficulty calculation.
-#[derive(Clone, Debug, Eq, Error, PartialEq)]
-pub enum AdjustedDifficultyError {
-    /// A full block did not expose its consensus height.
-    #[error("candidate block has no coinbase height")]
-    MissingCoinbaseHeight,
-    /// The caller requested contextual validation for genesis.
-    #[error("genesis has no predecessor difficulty context")]
-    Genesis,
-    /// The predecessor height could not produce a candidate height.
-    #[error("candidate height overflows the block-height range")]
-    HeightOverflow,
-    /// The context did not contain exactly the height-dependent predecessor span.
-    #[error("difficulty context has {actual} entries, expected exactly {expected}")]
-    ContextLength {
-        /// Required predecessor count.
-        expected: usize,
-        /// Supplied predecessor count, capped at one beyond the maximum span.
-        actual: usize,
-    },
-}
-
-impl AdjustedDifficulty {
-    /// Create an `AdjustedDifficulty` from a `candidate_block`, `network`, and `context`.
-    ///
-    /// The caller supplies the previous
-    /// `PoWAveragingWindow + PoWMedianBlockSpan` (28) `difficulty_threshold`s and
-    /// `time`s from the relevant chain for `candidate_block`, in reverse height
-    /// order, starting with the previous block.
-    ///
-    /// Miners supply block times.
-    /// The `time` values might not follow reverse chronological order.
-    pub fn new_from_block<C>(
-        candidate_block: &Block,
-        network: &Network,
-        context: C,
-    ) -> Result<AdjustedDifficulty, AdjustedDifficultyError>
-    where
-        C: IntoIterator<Item = (CompactDifficulty, DateTime<Utc>)>,
-    {
-        let candidate_block_height = candidate_block
-            .coinbase_height()
-            .ok_or(AdjustedDifficultyError::MissingCoinbaseHeight)?;
-        let previous_block_height =
-            (candidate_block_height - 1).ok_or(AdjustedDifficultyError::Genesis)?;
-
-        AdjustedDifficulty::new_from_header_time(
-            candidate_block.header.time,
-            previous_block_height,
-            network,
-            context,
-        )
-    }
-
-    /// Create an [`AdjustedDifficulty`] from header time, parent height, network, and context.
-    ///
-    /// Header validation uses this constructor before the node downloads the full block.
-    ///
-    /// See [`Self::new_from_block`] for detailed information about the `context`.
-    ///
-    pub fn new_from_header_time<C>(
-        candidate_header_time: DateTime<Utc>,
-        previous_block_height: block::Height,
-        network: &Network,
-        context: C,
-    ) -> Result<AdjustedDifficulty, AdjustedDifficultyError>
-    where
-        C: IntoIterator<Item = (CompactDifficulty, DateTime<Utc>)>,
-    {
-        let candidate_height =
-            (previous_block_height + 1).ok_or(AdjustedDifficultyError::HeightOverflow)?;
-
-        let (thresholds, times) = context
-            .into_iter()
-            .take(POW_ADJUSTMENT_BLOCK_SPAN + 1)
-            .unzip::<_, _, Vec<_>, Vec<_>>();
-
-        let span = u32::try_from(POW_ADJUSTMENT_BLOCK_SPAN)
-            .map_err(|_| AdjustedDifficultyError::HeightOverflow)?;
-        let expected = usize::try_from(candidate_height.0.min(span))
-            .map_err(|_| AdjustedDifficultyError::HeightOverflow)?;
-        if thresholds.len() != expected {
-            return Err(AdjustedDifficultyError::ContextLength {
-                expected,
-                actual: thresholds.len(),
-            });
-        }
-
-        let actual = thresholds.len();
-        let relevant_difficulty_thresholds: BoundedVec<
-            CompactDifficulty,
-            1,
-            POW_ADJUSTMENT_BLOCK_SPAN,
-        > = thresholds
-            .try_into()
-            .map_err(|_| AdjustedDifficultyError::ContextLength { expected, actual })?;
-        let relevant_times: BoundedVec<DateTime<Utc>, 1, POW_ADJUSTMENT_BLOCK_SPAN> = times
-            .try_into()
-            .map_err(|_| AdjustedDifficultyError::ContextLength { expected, actual })?;
-
-        Ok(AdjustedDifficulty {
-            candidate_time: candidate_header_time,
-            candidate_height,
-            network: network.clone(),
-            relevant_difficulty_thresholds,
-            relevant_times,
-        })
-    }
-
-    /// Returns the candidate block's height.
-    pub fn candidate_height(&self) -> block::Height {
-        self.candidate_height
-    }
-
-    /// Returns the candidate block's time field.
-    pub fn candidate_time(&self) -> DateTime<Utc> {
-        self.candidate_time
-    }
-
-    /// Returns the configured network.
-    pub fn network(&self) -> Network {
-        self.network.clone()
-    }
-
-    /// Calculate the expected `difficulty_threshold` from the candidate block's time and height,
-    /// the network, and the
-    /// `difficulty_threshold`s and `time`s from the previous
-    /// `PoWAveragingWindow + PoWMedianBlockSpan` (28) blocks in the relevant chain.
-    ///
-    /// The difficulty calculation implements `ThresholdBits` from the Zcash specification and the Testnet
-    /// minimum difficulty adjustment from ZIPs 205 and 208.
-    pub fn expected_difficulty_threshold(&self) -> CompactDifficulty {
-        if NetworkUpgrade::is_testnet_min_difficulty_block(
-            &self.network,
-            self.candidate_height,
-            self.candidate_time,
-            *self.relevant_times.first(),
-        ) {
-            assert!(
-                self.network.is_a_test_network(),
-                "invalid network: the minimum difficulty rule only applies on test networks"
-            );
-            self.network.target_difficulty_limit().to_compact()
-        } else {
-            self.threshold_bits()
-        }
-    }
-
-    /// Calculate a candidate block's `difficulty_threshold` from its height, network, and context.
-    ///
-    /// See [`Self::expected_difficulty_threshold`] for details.
-    ///
-    /// The difficulty calculation implements `ThresholdBits` from the Zcash specification.
-    /// `ThresholdBits` excludes the Testnet minimum difficulty adjustment.
-    fn threshold_bits(&self) -> CompactDifficulty {
-        let averaging_window_height = u32::try_from(POW_AVERAGING_WINDOW)
-            .expect("averaging window is much smaller than u32::MAX");
-
-        if self.candidate_height.0 <= averaging_window_height {
-            // # Consensus
-            //
-            // `ThresholdBits(height)` is `PoWLimit` for `height <= PoWAveragingWindow`.
-            // Zakura starts full-block contextual validation after the mandatory checkpoint on
-            // Mainnet and Testnet. Only header sync and non-checkpointed test networks reach this
-            // early-chain path.
-            return self.network.target_difficulty_limit().to_compact();
-        }
-
-        let averaging_window_timespan = NetworkUpgrade::averaging_window_timespan_for_height(
-            &self.network,
-            self.candidate_height,
-        );
-
-        let threshold = (self.mean_target_difficulty() / averaging_window_timespan.num_seconds())
-            * self.median_timespan_bounded().num_seconds();
-        let threshold = min(self.network.target_difficulty_limit(), threshold);
-
-        threshold.to_compact()
-    }
-
-    /// Calculate the arithmetic mean of the expanded `difficulty_threshold` values from the
-    /// previous `PoWAveragingWindow` blocks in the relevant chain.
-    ///
-    /// Implements `MeanTarget` from the Zcash specification.
-    fn mean_target_difficulty(&self) -> ExpandedDifficulty {
-        // `threshold_bits` returns `PoWLimit` before it calls this function at early-chain heights.
-        // A valid relevant chain contains at least 17 blocks at later heights.
-
-        let averaging_window_thresholds =
-            &self.relevant_difficulty_thresholds.as_slice()[0..POW_AVERAGING_WINDOW];
-
-        let divisor: U256 = POW_AVERAGING_WINDOW.into();
-        let mut quotient_total = U256::zero();
-        let mut remainder_total = U256::zero();
-        for compact in averaging_window_thresholds {
-            let target: U256 = compact
-                .to_expanded()
-                .expect("difficulty thresholds in previously verified blocks are valid")
-                .into();
-            quotient_total = quotient_total
-                .checked_add(target / divisor)
-                .expect("the sum of divided targets is at most U256::MAX");
-            remainder_total = remainder_total
-                .checked_add(target % divisor)
-                .expect("17 remainders smaller than 17 fit in U256");
-        }
-        ExpandedDifficulty::from(
-            quotient_total
-                .checked_add(remainder_total / divisor)
-                .expect("the exact mean of U256 targets is at most U256::MAX"),
-        )
-    }
-
-    /// Calculate the bounded median timespan.
-    /// The calculation subtracts the medians of the `time` values from
-    /// the previous `PoWAveragingWindow + PoWMedianBlockSpan` (28) blocks in the
-    /// relevant chain.
-    ///
-    /// The difficulty calculation uses the candidate block's height and network to calculate the
-    /// `AveragingWindowTimespan` for that block.
-    ///
-    /// `PoWDampingFactor` damps the median timespan.
-    /// `PoWMaxAdjustDown` and `PoWMaxAdjustUp` bound the median timespan.
-    ///
-    /// Implements `ActualTimespanBounded` from the Zcash specification.
-    ///
-    /// The calculation uses only `PoWMedianBlockSpan` times at each end of the timespan.
-    /// The calculation ignores times `11..=16`.
-    fn median_timespan_bounded(&self) -> Duration {
-        let averaging_window_timespan = NetworkUpgrade::averaging_window_timespan_for_height(
-            &self.network,
-            self.candidate_height,
-        );
-        // The duration value is exact. The calculation must truncate its nanoseconds component.
-        let damped_variance =
-            (self.median_timespan() - averaging_window_timespan) / POW_DAMPING_FACTOR;
-        // `num_seconds` truncates negative values toward zero as the Zcash specification requires.
-        let damped_variance = Duration::seconds(damped_variance.num_seconds());
-
-        // `ActualTimespanDamped` in the Zcash specification
-        let median_timespan_damped = averaging_window_timespan + damped_variance;
-
-        // `MinActualTimespan` and `MaxActualTimespan` in the Zcash spec
-        let min_median_timespan =
-            averaging_window_timespan * (100 - POW_MAX_ADJUST_UP_PERCENT) / 100;
-        let max_median_timespan =
-            averaging_window_timespan * (100 + POW_MAX_ADJUST_DOWN_PERCENT) / 100;
-
-        // `ActualTimespanBounded` in the Zcash specification
-        max(
-            min_median_timespan,
-            min(max_median_timespan, median_timespan_damped),
-        )
-    }
-
-    /// Calculate the median timespan.
-    /// The calculation subtracts the medians of the `time` values from
-    /// `PoWAveragingWindow + PoWMedianBlockSpan` (28) blocks in the relevant chain.
-    ///
-    /// Implements `ActualTimespan` from the Zcash specification.
-    ///
-    /// See [`Self::median_timespan_bounded`] for details.
-    fn median_timespan(&self) -> Duration {
-        let newer_median = self.median_time_past();
-
-        // MedianTime(height : N) := median([ nTime(𝑖) for 𝑖 from max(0, height − PoWMedianBlockSpan) up to max(0, height − 1) ])
-        let older_median = if self.relevant_times.len() > POW_AVERAGING_WINDOW {
-            let older_times: Vec<_> = self
-                .relevant_times
-                .iter()
-                .skip(POW_AVERAGING_WINDOW)
-                .cloned()
-                .take(POW_MEDIAN_BLOCK_SPAN)
-                .collect();
-
-            AdjustedDifficulty::median_time(older_times)
-        } else {
-            *self.relevant_times.last()
-        };
-
-        // `ActualTimespan` in the Zcash specification
-        newer_median - older_median
-    }
-
-    /// Calculate the median of the `time`s from the previous
-    /// `PoWMedianBlockSpan` (11) blocks in the relevant chain.
-    ///
-    /// The median-time calculation implements `median-time-past` and `MedianTime(candidate_height)` from the
-    /// Zcash specification. Both specification functions produce the same result.
-    pub fn median_time_past(&self) -> DateTime<Utc> {
-        let median_times: Vec<DateTime<Utc>> = self
-            .relevant_times
-            .iter()
-            .take(POW_MEDIAN_BLOCK_SPAN)
-            .cloned()
-            .collect();
-
-        AdjustedDifficulty::median_time(median_times)
-    }
-
-    /// Calculate the median of the `median_block_span_times`: the `time`s from a
-    /// Vec of `PoWMedianBlockSpan` (11) or fewer blocks in the relevant chain.
-    ///
-    /// Implements `MedianTime` from the Zcash specification.
-    ///
-    /// # Panics
-    ///
-    /// The median-time calculation panics if the caller provides an empty `Vec`.
-    pub fn median_time(mut median_block_span_times: Vec<DateTime<Utc>>) -> DateTime<Utc> {
-        median_block_span_times.sort_unstable();
-
-        // > median(𝑆) := sorted(𝑆)_{ceiling((length(𝑆)+1)/2)}
-        // <https://zips.z.cash/protocol/protocol.pdf>, section 7.7.3, Difficulty Adjustment (p. 132)
-        let median_idx = median_block_span_times.len() / 2;
-        median_block_span_times[median_idx]
-    }
-}
-
-/// Contextual candidate time or difficulty failure.
-#[derive(Clone, Debug, Eq, Error, PartialEq)]
-pub enum ContextualValidationError {
-    /// Candidate time is not strictly greater than median-time-past.
-    #[error("block time {candidate_time:?} is less than or equal to median-time-past {median_time_past:?}")]
-    TimeTooEarly {
-        /// Candidate header time.
-        candidate_time: DateTime<Utc>,
-        /// Median of the preceding header times.
-        median_time_past: DateTime<Utc>,
-    },
-    /// Candidate time exceeds the active 90-minute median-time-past limit.
-    #[error("block time {candidate_time:?} exceeds maximum {block_time_max:?}")]
-    TimeTooLate {
-        /// Candidate header time.
-        candidate_time: DateTime<Utc>,
-        /// Inclusive maximum candidate time.
-        block_time_max: DateTime<Utc>,
-    },
-    /// Candidate compact target does not match the contextual expected target.
-    #[error(
-        "block difficulty {difficulty_threshold:?} does not match expected {expected_difficulty:?}"
-    )]
-    InvalidDifficultyThreshold {
-        /// Candidate header compact target.
-        difficulty_threshold: CompactDifficulty,
-        /// Compact target calculated from the branch-local context.
-        expected_difficulty: CompactDifficulty,
-    },
-}
-
-/// Validate contextual median-time and compact-target rules using exact branch-local context.
-pub fn validate_contextual_difficulty_and_time(
-    difficulty_threshold: CompactDifficulty,
-    difficulty_adjustment: AdjustedDifficulty,
-) -> Result<(), ContextualValidationError> {
-    let candidate_height = difficulty_adjustment.candidate_height();
-    let candidate_time = difficulty_adjustment.candidate_time();
-    let network = difficulty_adjustment.network();
-    let median_time_past = difficulty_adjustment.median_time_past();
-    let block_time_max = median_time_past + Duration::seconds(BLOCK_MAX_TIME_SINCE_MEDIAN.into());
-
-    let genesis_height = NetworkUpgrade::Genesis
-        .activation_height(&network)
-        .expect("Zakura always has a genesis height available");
-
-    if candidate_time <= median_time_past && candidate_height != genesis_height {
-        return Err(ContextualValidationError::TimeTooEarly {
-            candidate_time,
-            median_time_past,
-        });
-    }
-
-    // This rule excludes Mainnet height 1 and Testnet heights below 653,606.
-    if candidate_height.0 >= 2
-        && network.is_max_block_time_enforced(candidate_height)
-        && candidate_time > block_time_max
-    {
-        return Err(ContextualValidationError::TimeTooLate {
-            candidate_time,
-            block_time_max,
-        });
-    }
-
-    let expected_difficulty = difficulty_adjustment.expected_difficulty_threshold();
-    if network.disable_pow() {
-        if difficulty_threshold.to_work().is_none() {
-            return Err(ContextualValidationError::InvalidDifficultyThreshold {
-                difficulty_threshold,
-                expected_difficulty,
-            });
-        }
-    } else if difficulty_threshold != expected_difficulty {
-        return Err(ContextualValidationError::InvalidDifficultyThreshold {
-            difficulty_threshold,
-            expected_difficulty,
-        });
-    }
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{hint::black_box, time::Instant};
-
-    use super::*;
-    use zakura_chain::parameters::testnet::{Parameters, RegtestParameters};
-
-    fn compact_half_limit(network: &Network) -> CompactDifficulty {
-        (network.target_difficulty_limit() / U256::from(2_u8)).to_compact()
-    }
-
-    #[test]
-    fn custom_target_mean_does_not_overflow_u256() {
-        let candidate_time =
-            DateTime::from_timestamp(2_000_000_000, 0).expect("test timestamp is in range");
-        let compact = ExpandedDifficulty::from(U256::MAX).to_compact();
-        let expanded = compact
-            .to_expanded()
-            .expect("the maximum compact-representable target is valid");
-        let context = vec![(compact, candidate_time - Duration::seconds(1)); 28];
-        let adjustment = AdjustedDifficulty::new_from_header_time(
-            candidate_time,
-            block::Height(699_999),
-            &Network::Mainnet,
-            context,
-        )
-        .expect("the complete context is accepted");
-
-        assert_eq!(adjustment.mean_target_difficulty(), expanded);
-    }
-
-    fn context(
-        network: &Network,
-        candidate_time: DateTime<Utc>,
-        spacing: Duration,
-        len: usize,
-    ) -> Vec<(CompactDifficulty, DateTime<Utc>)> {
-        let difficulty = compact_half_limit(network);
-        (1..=len)
-            .map(|offset| {
-                let offset = i32::try_from(offset).expect("test context length fits in i32");
-                (difficulty, candidate_time - spacing * offset)
-            })
-            .collect()
-    }
-
-    fn validate_with_expected_target(
-        network: &Network,
-        candidate_height: block::Height,
-        candidate_time: DateTime<Utc>,
-        context: &[(CompactDifficulty, DateTime<Utc>)],
-    ) -> Result<(), ContextualValidationError> {
-        let previous_height = (candidate_height - 1).expect("test candidate is not genesis");
-        let adjustment = match AdjustedDifficulty::new_from_header_time(
-            candidate_time,
-            previous_height,
-            network,
-            context.iter().copied(),
-        ) {
-            Ok(adjustment) => adjustment,
-            Err(error) => panic!("the helper requires an exact height-dependent context: {error}"),
-        };
-        let expected = adjustment.expected_difficulty_threshold();
-        validate_contextual_difficulty_and_time(expected, adjustment)
-    }
-
-    #[test]
-    fn difficulty_windows_upgrades_testnet_minimum_and_partitions_match() {
-        let candidate_time =
-            DateTime::from_timestamp(2_000_000_000, 0).expect("test timestamp is in range");
-
-        for network in Network::iter() {
-            for len in [1, 11, 16, 17, 27] {
-                let candidate_height = block::Height(700_000);
-                let spacing = NetworkUpgrade::target_spacing_for_height(&network, candidate_height);
-                let context = context(&network, candidate_time, spacing, len);
-                let previous_height = (candidate_height - 1).expect("height is positive");
-                assert!(
-                    matches!(
-                        AdjustedDifficulty::new_from_header_time(
-                            candidate_time,
-                            previous_height,
-                            &network,
-                            context,
-                        ),
-                        Err(AdjustedDifficultyError::ContextLength {
-                            expected: POW_ADJUSTMENT_BLOCK_SPAN,
-                            actual,
-                        }) if actual == len
-                    ),
-                    "late-chain contexts must not fail open for {network:?}"
-                );
-            }
-
-            let candidate_height = block::Height(700_000);
-            let spacing = NetworkUpgrade::target_spacing_for_height(&network, candidate_height);
-            let late_context =
-                context(&network, candidate_time, spacing, POW_ADJUSTMENT_BLOCK_SPAN);
-            let previous_height = (candidate_height - 1).expect("height is positive");
-            let expected = AdjustedDifficulty::new_from_header_time(
-                candidate_time,
-                previous_height,
-                &network,
-                late_context.iter().copied(),
-            )
-            .expect("the late-chain context contains the complete span")
-            .expected_difficulty_threshold();
-            for split in 0..=late_context.len() {
-                let partitioned = late_context[..split]
-                    .iter()
-                    .chain(&late_context[split..])
-                    .copied();
-                assert_eq!(
-                    AdjustedDifficulty::new_from_header_time(
-                        candidate_time,
-                        previous_height,
-                        &network,
-                        partitioned,
-                    )
-                    .expect("partitioning preserves the complete context")
-                    .expected_difficulty_threshold(),
-                    expected,
-                    "response partitions must not affect difficulty for {network:?}, split {split}"
-                );
-            }
-
-            for (height, _) in network.activation_list() {
-                if height == block::Height(0) {
-                    continue;
-                }
-                let spacing = NetworkUpgrade::target_spacing_for_height(&network, height);
-                let context_len = usize::try_from(height.0.min(28))
-                    .expect("bounded test context length fits in usize");
-                let context = context(&network, candidate_time, spacing, context_len);
-                validate_with_expected_target(&network, height, candidate_time, &context)
-                    .expect("the shared result validates at every configured upgrade boundary");
-            }
-        }
-
-        let testnet = Network::new_default_testnet();
-        let activation_height = block::Height(299_188);
-        let spacing = NetworkUpgrade::target_spacing_for_height(&testnet, activation_height);
-        let previous_time = candidate_time - spacing * 6;
-        let mut exact_gap = context(&testnet, candidate_time, spacing, 28);
-        exact_gap[0].1 = previous_time;
-        let previous_height = (activation_height - 1).expect("height is positive");
-        let exact_gap_target = AdjustedDifficulty::new_from_header_time(
-            candidate_time,
-            previous_height,
-            &testnet,
-            exact_gap,
-        )
-        .expect("the test supplies the complete late-chain context")
-        .expected_difficulty_threshold();
-        assert_ne!(
-            exact_gap_target,
-            testnet.target_difficulty_limit().to_compact()
-        );
-
-        let minimum_time = candidate_time + Duration::seconds(1);
-        let minimum_context = context(&testnet, minimum_time, spacing, 28)
-            .into_iter()
-            .enumerate()
-            .map(|(index, (difficulty, time))| {
-                if index == 0 {
-                    (difficulty, previous_time)
-                } else {
-                    (difficulty, time)
-                }
-            });
-        assert_eq!(
-            AdjustedDifficulty::new_from_header_time(
-                minimum_time,
-                previous_height,
-                &testnet,
-                minimum_context,
-            )
-            .expect("the test supplies the complete late-chain context")
-            .expected_difficulty_threshold(),
-            testnet.target_difficulty_limit().to_compact(),
-            "ZIP 205/208 minimum difficulty begins strictly above six target spacings"
-        );
-    }
-
-    #[test]
-    fn difficulty_damping_bounds_are_exact() {
-        let network = Network::Mainnet;
-        let candidate_height = block::Height(700_000);
-        let previous_height = (candidate_height - 1).expect("height is positive");
-        let candidate_time =
-            DateTime::from_timestamp(2_000_000_000, 0).expect("test timestamp is in range");
-        let averaging_timespan =
-            NetworkUpgrade::averaging_window_timespan_for_height(&network, candidate_height);
-        let mean_target = compact_half_limit(&network)
-            .to_expanded()
-            .expect("the test target is valid");
-
-        let fast_context = context(
-            &network,
-            candidate_time,
-            Duration::seconds(1),
-            POW_ADJUSTMENT_BLOCK_SPAN,
-        );
-        let fast = AdjustedDifficulty::new_from_header_time(
-            candidate_time,
-            previous_height,
-            &network,
-            fast_context,
-        )
-        .expect("the test supplies the complete late-chain context")
-        .expected_difficulty_threshold();
-        let minimum_timespan = averaging_timespan * (100 - POW_MAX_ADJUST_UP_PERCENT) / 100;
-        assert_eq!(
-            fast,
-            ((mean_target / averaging_timespan.num_seconds()) * minimum_timespan.num_seconds())
-                .to_compact(),
-            "fast blocks are clipped at the 16% upward-adjustment bound"
-        );
-
-        let slow_context = context(
-            &network,
-            candidate_time,
-            Duration::seconds(10_000),
-            POW_ADJUSTMENT_BLOCK_SPAN,
-        );
-        let slow = AdjustedDifficulty::new_from_header_time(
-            candidate_time,
-            previous_height,
-            &network,
-            slow_context,
-        )
-        .expect("the test supplies the complete late-chain context")
-        .expected_difficulty_threshold();
-        let maximum_timespan = averaging_timespan * (100 + POW_MAX_ADJUST_DOWN_PERCENT) / 100;
-        assert_eq!(
-            slow,
-            ((mean_target / averaging_timespan.num_seconds()) * maximum_timespan.num_seconds())
-                .to_compact(),
-            "slow blocks are clipped at the 32% downward-adjustment bound"
-        );
-    }
-
-    #[test]
-    fn median_and_production_max_time_boundaries_are_exact() {
-        let base = DateTime::from_timestamp(1_600_000_000, 0).expect("test timestamp is in range");
-        let difficulty = Network::Mainnet.target_difficulty_limit().to_compact();
-
-        for len in 1..=POW_ADJUSTMENT_BLOCK_SPAN {
-            let times: Vec<_> = (0..len)
-                .map(|offset| base + Duration::seconds(i64::try_from(offset).expect("fits")))
-                .rev()
-                .collect();
-            let adjustment = AdjustedDifficulty::new_from_header_time(
-                base + Duration::hours(1),
-                block::Height(
-                    u32::try_from(len - 1).expect("the bounded context length fits in u32"),
-                ),
-                &Network::Mainnet,
-                times.iter().copied().map(|time| (difficulty, time)),
-            )
-            .expect("early-chain context length equals the candidate height");
-            let mut expected: Vec<_> = times.into_iter().take(POW_MEDIAN_BLOCK_SPAN).collect();
-            expected.sort_unstable();
-            assert_eq!(adjustment.median_time_past(), expected[expected.len() / 2]);
-        }
-
-        for (network, height, max_is_active) in [
-            (Network::Mainnet, block::Height(1), false),
-            (Network::Mainnet, block::Height(2), true),
-            (
-                Network::new_default_testnet(),
-                block::Height(653_605),
-                false,
-            ),
-            (Network::new_default_testnet(), block::Height(653_606), true),
-        ] {
-            let context = vec![
-                (network.target_difficulty_limit().to_compact(), base);
-                usize::try_from(height.0.min(28))
-                    .expect("bounded height fits in usize")
-            ];
-            assert!(matches!(
-                validate_with_expected_target(&network, height, base, &context),
-                Err(ContextualValidationError::TimeTooEarly { .. })
-            ));
-
-            let equality = base + Duration::minutes(90);
-            validate_with_expected_target(&network, height, equality, &context)
-                .expect("the 90-minute equality boundary is inclusive");
-
-            let one_second_above = equality + Duration::seconds(1);
-            let result =
-                validate_with_expected_target(&network, height, one_second_above, &context);
-            assert_eq!(
-                matches!(result, Err(ContextualValidationError::TimeTooLate { .. })),
-                max_is_active,
-                "unexpected max-time activation for {network:?} at {height:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn disabled_pow_never_waives_median_time() {
-        let network = Network::new_regtest(RegtestParameters::default());
-        assert!(network.disable_pow());
-        let time = DateTime::from_timestamp(1_700_000_000, 0).expect("test timestamp is in range");
-        let context = [(network.target_difficulty_limit().to_compact(), time)];
-        let adjustment =
-            AdjustedDifficulty::new_from_header_time(time, block::Height(0), &network, context)
-                .expect("height one requires exactly one predecessor");
-        assert!(matches!(
-            validate_contextual_difficulty_and_time(
-                network.target_difficulty_limit().to_compact(),
-                adjustment,
-            ),
-            Err(ContextualValidationError::TimeTooEarly { .. })
-        ));
-    }
-
-    #[test]
-    fn custom_mtp_and_max_time_use_local_parameters_with_pow_on_or_off() {
-        let base = DateTime::from_timestamp(1_700_000_000, 0).expect("test timestamp is in range");
-        let activation = block::Height(10);
-
-        for disable_pow in [false, true] {
-            let network = Parameters::build()
-                .with_network_name(if disable_pow {
-                    "CustomTimePowOff"
-                } else {
-                    "CustomTimePowOn"
-                })
-                .expect("the custom network name is valid")
-                .with_disable_pow(disable_pow)
-                .with_max_block_time_start_height(activation)
-                .to_network()
-                .expect("the custom-network parameters are valid");
-            let context_before = vec![(network.target_difficulty_limit().to_compact(), base); 9];
-            let context_at = vec![(network.target_difficulty_limit().to_compact(), base); 10];
-
-            assert!(matches!(
-                validate_with_expected_target(&network, block::Height(9), base, &context_before),
-                Err(ContextualValidationError::TimeTooEarly { .. })
-            ));
-            assert!(
-                validate_with_expected_target(
-                    &network,
-                    block::Height(9),
-                    base + Duration::minutes(90) + Duration::seconds(1),
-                    &context_before,
-                )
-                .is_ok(),
-                "the local maximum-time rule is not active before its configured height"
-            );
-            assert!(matches!(
-                validate_with_expected_target(
-                    &network,
-                    activation,
-                    base + Duration::minutes(90) + Duration::seconds(1),
-                    &context_at,
-                ),
-                Err(ContextualValidationError::TimeTooLate { .. })
-            ));
-        }
-
-        let regtest = Network::new_regtest(RegtestParameters::default());
-        assert!(
-            !regtest.is_max_block_time_enforced(block::Height(1))
-                && regtest.is_max_block_time_enforced(block::Height(2)),
-            "Regtest must use its local policy rather than public Testnet height 653,606"
-        );
-    }
-
-    #[allow(clippy::unwrap_in_result)]
-    fn validate_contextual_sequence(
-        network: &Network,
-        mut parent_height: block::Height,
-        predecessors: &[(CompactDifficulty, DateTime<Utc>)],
-        candidates: &[(CompactDifficulty, DateTime<Utc>)],
-    ) -> Result<(), ContextualValidationError> {
-        let mut context = predecessors.to_vec();
-        for (difficulty, time) in candidates {
-            validate_contextual_difficulty_and_time(
-                *difficulty,
-                AdjustedDifficulty::new_from_header_time(
-                    *time,
-                    parent_height,
-                    network,
-                    context.iter().copied(),
-                )
-                .expect("the benchmark retains the complete late-chain context"),
-            )?;
-            parent_height = parent_height
-                .next()
-                .expect("the benchmark range is far below the height limit");
-            context.insert(0, (*difficulty, *time));
-            context.truncate(POW_ADJUSTMENT_BLOCK_SPAN);
-        }
-        Ok(())
-    }
-
-    /// Run with:
-    /// `cargo test -p zakura-header-chain --release -- --ignored --nocapture contextual_writer_hold_microbench`
-    #[test]
-    #[ignore = "manual release-mode benchmark for the R5 writer-boundary gate"]
-    #[allow(clippy::print_stdout)]
-    fn contextual_writer_hold_microbench() {
-        const TYPICAL_HEADERS: usize = 32;
-        const MAX_HEADERS: usize = 4_096;
-        const TYPICAL_ITERATIONS: u32 = 2_000;
-        const MAX_ITERATIONS: u32 = 20;
-        const MAX_ACCEPTABLE_WRITER_HOLD: std::time::Duration =
-            std::time::Duration::from_millis(25);
-
-        let network = Network::new_regtest(RegtestParameters::default());
-        let parent_height = block::Height(700_000);
-        let base =
-            DateTime::from_timestamp(2_000_000_000, 0).expect("benchmark timestamp is in range");
-        let spacing = Duration::seconds(75);
-        let difficulty = network.target_difficulty_limit().to_compact();
-        let predecessors = context(&network, base, spacing, POW_ADJUSTMENT_BLOCK_SPAN);
-        let candidates = |count: usize| {
-            (1..=count)
-                .map(|offset| {
-                    let offset =
-                        i32::try_from(offset).expect("the maximum batch length fits in an i32");
-                    (difficulty, base + spacing * offset)
-                })
-                .collect::<Vec<_>>()
-        };
-        let typical = candidates(TYPICAL_HEADERS);
-        let maximum = candidates(MAX_HEADERS);
-
-        println!(
-            "R5 contextual writer-hold gate: max 4,096-header average < {:?}",
-            MAX_ACCEPTABLE_WRITER_HOLD
-        );
-        for mode in ["integrated", "headers-only"] {
-            for (case, batch, iterations) in [
-                ("typical-pass", typical.as_slice(), TYPICAL_ITERATIONS),
-                ("maximum-pass", maximum.as_slice(), MAX_ITERATIONS),
-            ] {
-                let started = Instant::now();
-                for _ in 0..iterations {
-                    validate_contextual_sequence(
-                        &network,
-                        parent_height,
-                        &predecessors,
-                        black_box(batch),
-                    )
-                    .expect("the benchmark sequence is contextually valid");
-                }
-                let average = started.elapsed() / iterations;
-                println!(
-                    "mode={mode} case={case} headers={} average={average:?} per_header={:?}",
-                    batch.len(),
-                    average / u32::try_from(batch.len()).expect("benchmark length fits in u32"),
-                );
-                if case == "maximum-pass" {
-                    assert!(
-                        average < MAX_ACCEPTABLE_WRITER_HOLD,
-                        "{mode} maximum batch held the prototype writer gate for {average:?}"
-                    );
-                }
-            }
-
-            for (case, invalid_offset) in [("invalid-first", 0), ("invalid-last", MAX_HEADERS - 1)]
-            {
-                let mut invalid = maximum.clone();
-                invalid[invalid_offset].1 = predecessors[5].1;
-                let started = Instant::now();
-                let result = validate_contextual_sequence(
-                    &network,
-                    parent_height,
-                    &predecessors,
-                    black_box(&invalid),
-                );
-                let elapsed = started.elapsed();
-                assert!(result.is_err(), "{case} must exercise contextual rejection");
-                println!(
-                    "mode={mode} case={case} headers_examined={} elapsed={elapsed:?}",
-                    invalid_offset + 1,
-                );
-            }
-        }
-    }
-}
